+++ .env.example
@@ -1,0 +1,3 @@
+HTTP_PORT=3000
+WS_PORT=80
+SCREEN_SIZE=200
