+++ index.ts
@@ -1,0 +1,8 @@
+import { httpServerStart } from './src/http_server/index';
+import { wsServerStart } from './src/ws_server/index';
+
+httpServerStart();
+wsServerStart();
+
+
+
